Fixes for thousand separator and Contract Price Error

+++ CommanderContracts-iOS/Helpers/ClientOrUserSingleton.swift
@@ -89,9 +89,9 @@ class ClientOrUserSingleton {
     
     
     
-    func setContractAmount(theAmount: Int) -> () {
+    func setContractAmount(theAmount: Double) -> () {
         
-        contractAmount = Double(Int(theAmount))
+        contractAmount = theAmount
         
     }
